Use uint16 for byte position
Byte code expects 2 bytes exactly. Therefore use accurate data type to store byte pos

+++ generator/emit/il_code_generation_visitor.go
@@ -15,7 +15,7 @@ type ILCodeGenerationVisitor struct {
 	ilBuilder   *ILBuilder
 	function    *symbol.FunctionSymbol
 	assembler   *ILAssembler
-	bytePos     int
+	bytePos     uint16
 	Errors      []error
 }
